ARM: OMAP4: clock: Define 4470 flag in cpu masks
Define OMAP4470 flag in clock cpu masks.
Reuse CK_44XX clocks for 4470.

Increased 'flags' field size to fit additional CPU type.

+++ arch/arm/mach-omap2/clock.c
@@ -36,7 +36,7 @@
 #include "cm-regbits-24xx.h"
 #include "cm-regbits-34xx.h"
 
-u8 cpu_mask;
+u16 cpu_mask;
 
 /*
  * OMAP2+ specific clock functions
